Break out decorate function so internal sdk can still require it
