+++ lib/api.js
@@ -2,6 +2,7 @@
 
 var assign = require('object-assign');
 var Bluebird = require('bluebird');
+var decorate = require('./decorate');
 
 Bluebird.config({
     warnings: false,
@@ -10,25 +11,6 @@ Bluebird.config({
 
 if (typeof process === 'object' && process + '' === '[object process]') {
     global.window = {};
-}
-
-function decorateWithFor(originalInstance, factory, options) {
-    var instanceMap = {};
-
-    originalInstance.for = function (mappingKey) {
-        if (!mappingKey) {
-            return originalInstance;
-        }
-
-        if (!instanceMap[mappingKey]) {
-            // create a new instance
-            instanceMap[mappingKey] = factory(options);
-        }
-
-        return instanceMap[mappingKey];
-    };
-
-    return originalInstance;
 }
 
 var endpointFactories = {
@@ -54,7 +36,7 @@ function createEndpoints (options) {
       var factory = endpointFactories[endpointName];
       var originalInstance = factory(options);
 
-      endpoints[endpointName] = decorateWithFor(originalInstance, factory, options);
+      endpoints[endpointName] = decorate(originalInstance, factory, options);
   });
 
   return endpoints;
