+++ lib/api/profiles.js
@@ -76,7 +76,23 @@ module.exports = function profiles(options) {
              method: 'GET',
              resource: '/profiles?email={email}',
              args: ['email']
-         })
+         }),
+
+         /**
+          * Retrieve a profile by link
+          *
+          * @method
+          * @memberof api.profiles
+          * @param {string} link - Short name
+          * @returns {promise}
+          */
+          retrieveByLink: utils.requestFun({
+              authFlow: options.authFlow,
+              baseUrl: options.baseUrl,
+              method: 'GET',
+              resource: '/profiles?link={link}',
+              args: ['link']
+          })
 
     };
 };
